project/requirements.txt
aiofile==3.9.0
amazon-transcribe==0.6.2
awscrt==0.16.26
boto3==1.35.26
botocore==1.35.26
caio==0.9.17
contourpy==1.3.3
cycler==0.12.1
Cython==3.0.11
docutils==0.21.2
et_xmlfile==2.0.0
flask==3.1.2
flask-socketio==5.5.1
fonttools==4.60.1
jmespath==1.0.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.1.1
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pandas_flavor==0.7.0
patsy==1.0.2
pillow==10.4.0
pingouin==0.5.5
PyAudio==0.2.14
pyobjc-core==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyparsing==3.2.5
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.16.1
pyzmq==25.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.10.2
scikit-learn==1.7.2
scipy==1.16.3
screeninfo==0.8.1
seaborn==0.13.2
setuptools==78.1.0
six==1.16.0
statsmodels==0.14.5
tabulate==0.9.0
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.2.3
xarray==2025.10.1
msgpack==1.0.8
